+++ tests/testthat/test-test.metric.R
@@ -1,0 +1,6 @@
+# context("test.metric")
+
+# ## Test
+# test_that("test.metric works", {
+
+# })